requirements.txt
google-generativeai
Pillow
openpyxl
pytesseract
